Rename adjuster's TestUtils to AdjustTestUtils
