+++ adjuster/src/test/java/org/jboss/pnc/reqour/adjust/provider/SbtProviderTest.java
@@ -8,7 +8,7 @@ import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 import jakarta.inject.Inject;
 import org.assertj.core.data.MapEntry;
-import org.jboss.pnc.reqour.adjust.TestUtils;
+import org.jboss.pnc.reqour.adjust.AdjustTestUtils;
 import org.jboss.pnc.reqour.adjust.config.ReqourAdjusterConfig;
 import org.jboss.pnc.reqour.adjust.profiles.SbtAdjustProfile;
 import org.jboss.pnc.reqour.common.utils.IOUtils;
@@ -22,8 +22,8 @@ import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.jboss.pnc.reqour.adjust.TestUtils.assertSystemPropertiesContainExactly;
-import static org.jboss.pnc.reqour.adjust.TestUtils.assertSystemPropertyHasValuesSortedByPriority;
+import static org.jboss.pnc.reqour.adjust.AdjustTestUtils.assertSystemPropertiesContainExactly;
+import static org.jboss.pnc.reqour.adjust.AdjustTestUtils.assertSystemPropertyHasValuesSortedByPriority;
 
 @QuarkusTest
 @TestProfile(SbtAdjustProfile.class)
@@ -33,7 +33,7 @@ class SbtProviderTest {
     ReqourAdjusterConfig config;
 
     @Inject
-    TestUtils testUtils;
+    AdjustTestUtils adjustTestUtils;
 
     static Path workdir;
 
@@ -51,7 +51,7 @@ class SbtProviderTest {
     void prepareCommand_standardTemporaryBuildWithPersistentAlignmentPreference_generatedCommandIsCorrect() {
         SbtProvider provider = new SbtProvider(
                 config.adjust(),
-                testUtils.getAdjustRequest(Path.of("sbt-request.json")),
+                adjustTestUtils.getAdjustRequest(Path.of("sbt-request.json")),
                 workdir,
                 null,
                 null,
